requirements.txt
asgiref==3.8.1
attrs==24.2.0
black==24.8.0
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
exceptiongroup==1.2.2
filelock==3.15.4
flake8==7.1.1
identify==2.6.0
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.8.0
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
pytest==8.3.2
pytest-django==4.8.0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
uritemplate==4.1.1
virtualenv==20.26.3
